Update Rocket's display
