+++ src/pages/rockets.js
@@ -1,29 +1,24 @@
 import React from 'react';
-import { useDispatch } from 'react-redux';
-import { bookRockets, cancelRocketBooking } from '../redux/rockets';
-/* eslint-disable react/prop-types */
-const Rocket = (props) => {
-  const { rocket } = props;
-  const dispatch = useDispatch();
+import { useSelector } from 'react-redux';
+import Rocket from './Rocket';
 
-  const handleClick = () => {
-    if (rocket.reserved && rocket.reserved === true) {
-      dispatch(cancelRocketBooking(rocket.id));
-    } else {
-      dispatch(bookRockets(rocket.id));
-    }
-  };
+const Rockets = () => {
+  const rockets = useSelector((state) => state.rockets);
 
   return (
-    <div className="rocket d-flex ">
-      <img className="rocket-image" src={rocket.flickr_images} alt={`${rocket.rocket_name}`} />
-      <div className="rocket-details d-col-flex">
-        <h2>{rocket.rocket_name}</h2>
-        <p>{rocket.description}</p>
-        <button type="button" onClick={handleClick}>Reserve Rocket</button>
-      </div>
+    <div className="rockets d-col-flex">
+      {
+        rockets.length > 0
+          ? rockets.map((rocket) => (
+            <Rocket
+              key={rocket.id}
+              rocket={rocket}
+            />
+          ))
+          : ''
+      }
     </div>
   );
 };
 
-export default Rocket;
+export default Rockets;
